ai-service/requirements.txt
flask>=3.0
gunicorn>=22.0
openai>=1.40
python-dotenv>=1.0
